fix(lspconfig): use safer 0.11 compatibility check (#3265)
Previously only the minor version has been checked for equality, which
means it is brittle whenever we reach 0.12. Independently of when this
is expected to be removed, it's simply much safer to check if the
expected functionality is available by testing `vim.lsp.config` for nil.

Fixes 84f8170fdab2a9bb82f36fe76c68bca2241c50d8